update: rename `updates` collector to `update`
