+++ internal/collector/update/update_test.go
@@ -1,0 +1,12 @@
+package update_test
+
+import (
+	"testing"
+
+	"github.com/prometheus-community/windows_exporter/internal/collector/update"
+	"github.com/prometheus-community/windows_exporter/internal/testutils"
+)
+
+func BenchmarkCollector(b *testing.B) {
+	testutils.FuncBenchmarkCollector(b, "printer", update.NewWithFlags)
+}
