+++ pkg/collector/collector.go
@@ -55,7 +55,7 @@ import (
 	"github.com/prometheus-community/windows_exporter/internal/collector/textfile"
 	"github.com/prometheus-community/windows_exporter/internal/collector/thermalzone"
 	"github.com/prometheus-community/windows_exporter/internal/collector/time"
-	"github.com/prometheus-community/windows_exporter/internal/collector/updates"
+	"github.com/prometheus-community/windows_exporter/internal/collector/update"
 	"github.com/prometheus-community/windows_exporter/internal/collector/vmware"
 	v1 "github.com/prometheus-community/windows_exporter/internal/perfdata/v1"
 	"github.com/prometheus-community/windows_exporter/internal/types"
@@ -122,7 +122,7 @@ func NewWithConfig(config Config) *MetricCollectors {
 	collectors[textfile.Name] = textfile.New(&config.Textfile)
 	collectors[thermalzone.Name] = thermalzone.New(&config.ThermalZone)
 	collectors[time.Name] = time.New(&config.Time)
-	collectors[updates.Name] = updates.New(&config.Updates)
+	collectors[update.Name] = update.New(&config.Update)
 	collectors[vmware.Name] = vmware.New(&config.Vmware)
 
 	return New(collectors)
